+++ tests/test__tcp.py
@@ -4,109 +4,72 @@ import unittest
 
 from netports import tcp
 
-ALL_TCP = list(range(1, 65536))
+ALL = list(range(1, 65536))
 
 
+# noinspection DuplicatedCode
 class Test(unittest.TestCase):
     """unittest tcp.py"""
 
-    def test_valid__itcp(self):
-        """itcp()"""
-        for kwargs, req in [
-            ({}, []),
-            (dict(items=""), []),
-            (dict(items=[]), []),
-            (dict(items=1), [1]),
-            (dict(items="1"), [1]),
-            (dict(items=[1]), [1]),
-            (dict(items=[65535]), [65535]),
-            (dict(items=[5, 5, 1, 3, 4]), [1, 3, 4, 5]),
-            (dict(items="3-5,1,3-5,1"), [1, 3, 4, 5]),
-
-            (dict(items=ALL_TCP), ALL_TCP),
-            (dict(items=ALL_TCP, verbose=True), ALL_TCP),
-            (dict(items=ALL_TCP, verbose=False), [-1]),
-
-            (dict(all=True), ALL_TCP),
-            (dict(all=True, verbose=True), ALL_TCP),
-            (dict(all=True, verbose=False), [-1]),
-            (dict(items="1", all=True), ALL_TCP),
-            (dict(items="1", all=True, verbose=True), ALL_TCP),
-            (dict(items="1", all=True, verbose=False), [-1]),
-
-            (dict(items="-1", verbose=False), [-1]),
-            (dict(items=["-1"], verbose=False), [-1]),
-            (dict(items=["-1", "2"], verbose=False), [-1]),
-            (dict(items=["typo", "-1"], verbose=False), [-1]),
-            (dict(items=-1, verbose=False), [-1]),
-            (dict(items=[-1], verbose=False), [-1]),
-            (dict(items=[-1, 2], verbose=False), [-1]),
-            (dict(items=[-2, -1], verbose=False), [-1]),
+    def test_valid__itcp__stcp(self):
+        """itcp() stcp()"""
+        for kwargs, req, req_ in [
+            # ports
+            ({}, [], ""),
+            (dict(items=""), [], ""),
+            (dict(items=[]), [], ""),
+            (dict(items=1), [1], "1"),
+            (dict(items="1"), [1], "1"),
+            (dict(items=[1]), [1], "1"),
+            (dict(items=["1"]), [1], "1"),
+            (dict(items=[65535]), [65535], "65535"),
+            (dict(items=[5, 5, 1, 3, 4]), [1, 3, 4, 5], "1,3-5"),
+            (dict(items="3-5,1,3-5,1"), [1, 3, 4, 5], "1,3-5"),
+            # 1-65535
+            (dict(items="1-65535"), [-1], "1-65535"),
+            (dict(items=ALL), [-1], "1-65535"),
+            (dict(items=ALL, verbose=False), [-1], "1-65535"),
+            (dict(items=ALL, verbose=True), ALL, "1-65535"),
+            (dict(items="1-65535,1"), [-1], "1-65535"),
+            (dict(items=[*ALL, 1]), [-1], "1-65535"),
+            # -1
+            (dict(items=-1), [-1], "1-65535"),
+            (dict(items="-1"), [-1], "1-65535"),
+            (dict(items=[-1]), [-1], "1-65535"),
+            (dict(items=["-1"]), [-1], "1-65535"),
+            (dict(items=[-1, 2]), [-1], "1-65535"),
+            (dict(items=["-1", "2"]), [-1], "1-65535"),
+            (dict(items=[-1], verbose=False), [-1], "1-65535"),
+            # all
+            (dict(all=True), [-1], "1-65535"),
+            (dict(all=True, verbose=False), [-1], "1-65535"),
+            (dict(all=True, verbose=True), ALL, "1-65535"),
+            (dict(items="1", all=True), [-1], "1-65535"),
+            (dict(items="1", all=True, verbose=False), [-1], "1-65535"),
+            (dict(items="1", all=True, verbose=True), ALL, "1-65535"),
         ]:
             result = tcp.itcp(**kwargs)
             self.assertEqual(result, req, msg=f"{kwargs=}")
+            result_ = tcp.stcp(**kwargs)
+            self.assertEqual(result_, req_, msg=f"{kwargs=}")
 
-    def test_invalid__itcp(self):
-        """itcp()"""
-        for items, error in [
-            ("-1", ValueError),
-            ("0", ValueError),
-            ("65536", ValueError),
-            ([0], ValueError),
-            ([-1], ValueError),
-            ([65536], ValueError),
+    def test_invalid__itcp__stcp(self):
+        """itcp() stcp()"""
+        for kwargs, error in [
+            # ports
+            (dict(items=0), ValueError),
+            (dict(items="0"), ValueError),
+            (dict(items=[0]), ValueError),
+            (dict(items=65536), ValueError),
+            (dict(items="65536"), ValueError),
+            (dict(items=[65536]), ValueError),
+            # typo
+            (dict(items="typo"), ValueError),
         ]:
-            with self.assertRaises(error, msg=f"{items=}"):
-                tcp.itcp(items)
-
-    def test_valid__stcp(self):
-        """stcp()"""
-        for kwargs, req in [
-            ({}, ""),
-            (dict(items=""), ""),
-            (dict(items=[]), ""),
-            (dict(items="1"), "1"),
-            (dict(items=1), "1"),
-            (dict(items=[1]), "1"),
-            (dict(items=[65535]), "65535"),
-            (dict(items=[5, 5, 1, 3, 4]), "1,3-5"),
-            (dict(items="3-5,1,3-5,1"), "1,3-5"),
-
-            (dict(items=ALL_TCP), "1-65535"),
-            (dict(items=ALL_TCP, verbose=True), "1-65535"),
-            (dict(items=ALL_TCP, verbose=False), "1-65535"),
-
-            (dict(all=True), "1-65535"),
-            (dict(all=True, verbose=True), "1-65535"),
-            (dict(all=True, verbose=False), "1-65535"),
-            (dict(items="1", all=True), "1-65535"),
-            (dict(items="1", all=True, verbose=True), "1-65535"),
-            (dict(items="1", all=True, verbose=False), "1-65535"),
-
-            (dict(items="-1", verbose=False), "1-65535"),
-            (dict(items=["-1"], verbose=False), "1-65535"),
-            (dict(items=["-1", "2"], verbose=False), "1-65535"),
-            (dict(items=["typo", "-1"], verbose=False), "1-65535"),
-            (dict(items=-1, verbose=False), "1-65535"),
-            (dict(items=[-1], verbose=False), "1-65535"),
-            (dict(items=[-1, 2], verbose=False), "1-65535"),
-            (dict(items=[-2, -1], verbose=False), "1-65535"),
-        ]:
-            result = tcp.stcp(**kwargs)
-            self.assertEqual(result, req, msg=f"{kwargs=}")
-
-    def test_invalid__stcp(self):
-        """stcp()"""
-        for items, error in [
-            ("-1", ValueError),
-            ("0", ValueError),
-            ("65536", ValueError),
-            ([0], ValueError),
-            ([-1], ValueError),
-            ([65536], ValueError),
-        ]:
-            with self.assertRaises(error, msg=f"{items=}"):
-                tcp.stcp(items)
+            with self.assertRaises(error, msg=f"{kwargs=}"):
+                tcp.itcp(**kwargs)
+            with self.assertRaises(error, msg=f"{kwargs=}"):
+                tcp.stcp(**kwargs)
 
 
 if __name__ == "__main__":